+++ txpipe/__init__.py
@@ -18,4 +18,4 @@ from .covariance import TXFourierGaussianCovariance
 from .metacal_gcr_input import TXMetacalGCRInput
 from .diagnostics import TXDiagnostics
 from .rowe_stats import TXRoweStatistics
-from .field_center_extractor import TXExposureInfo
+from .exposure_info import TXExposureInfo
